packages/gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anzen-gateway"
version = "0.1.0"
description = "Anzen Safety Gateway - FastAPI + NeMo Guardrails + Presidio for PII detection and masking"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Anzen Team", email = "team@anzen.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.0.0",
    "nemoguardrails>=0.8.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "structlog>=23.2.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.12.0",
    "coverage>=7.3.0",
]

[project.scripts]
anzen-gateway = "anzen_gateway.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/anzen_gateway"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/config",
    "/README.md",
]

[dependency-groups]
dev = [
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]
